+++ src/main/java/nextstep/ladder/interfaces/Ladder.java
@@ -1,0 +1,9 @@
+package nextstep.ladder.interfaces;
+
+import java.util.List;
+
+public interface Ladder {
+    List<Lines> getLines();
+
+    Position result();
+}
