+++ src/main/java/nextstep/ladder/view/OutputView.java
@@ -2,9 +2,9 @@ package nextstep.ladder.view;
 
 import java.util.Map;
 import java.util.Map.Entry;
-import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.dto.CommaSeparatedResult;
 import nextstep.ladder.dto.ResultDto;
+import nextstep.ladder.interfaces.Ladder;
 import nextstep.ladder.interfaces.Lines;
 
 public class OutputView {
